Fixed Carousel on Property Show and Book Views
Previously, it was a hacked together method of hiding the carousel when only one image was present for a listing, and shoving in a placeholder when there were none. This caused the console to be filled up with errors because it was trying to target carousel elements which didn't exist.

Now, if there aren't enough images to make a carousel, the entire carousel element won't be made, resulting in no errors. If there are no images, a placeholder will be shown.